move svg element creation to service
